unit tests added(boss,cleaner,feedback).

+++ src/main/java/com/example/hospitalcleaner/business/service/impl/CleanerServiceImpl.java
@@ -1,14 +1,12 @@
 package com.example.hospitalcleaner.business.service.impl;
 
 import com.example.hospitalcleaner.business.core.mapper.CleanerMapper;
-import com.example.hospitalcleaner.business.core.results.DataResult;
-import com.example.hospitalcleaner.business.core.results.Result;
-import com.example.hospitalcleaner.business.core.results.SuccessDataResult;
-import com.example.hospitalcleaner.business.core.results.SuccessResult;
+import com.example.hospitalcleaner.business.core.results.*;
 import com.example.hospitalcleaner.business.dto.CleanerEntityDto;
 import com.example.hospitalcleaner.business.requests.CleanerEntityCRequest;
 import com.example.hospitalcleaner.business.requests.CleanerEntityDRequest;
 import com.example.hospitalcleaner.business.requests.CleanerEntityURequest;
+import com.example.hospitalcleaner.business.service.BossService;
 import com.example.hospitalcleaner.business.service.CleanerService;
 import com.example.hospitalcleaner.dataAccess.CleanerEntityRepository;
 import com.example.hospitalcleaner.entity.CleanerEntity;
@@ -16,12 +14,16 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
+
 @Service
 public class CleanerServiceImpl implements CleanerService {
     @Autowired
     private CleanerMapper cleanerMapper;
     @Autowired
     private CleanerEntityRepository cleanerEntityRepository;
+    @Autowired
+    private BossService bossService;
 
 
     @Override
@@ -31,6 +33,9 @@ public class CleanerServiceImpl implements CleanerService {
 
     @Override
     public Result add(CleanerEntityCRequest cleanerEntityCRequest) {
+        if(bossService.getById(cleanerEntityCRequest.getBossId()).getData()==null){
+            return new ErrorResult("boss bulunamadı.");
+        }
         CleanerEntity cleanerEntity=cleanerMapper.cleanerCreateToEntity(cleanerEntityCRequest);
         this.cleanerEntityRepository.save(cleanerEntity);
         return new SuccessResult("eklendi.");
@@ -38,18 +43,29 @@ public class CleanerServiceImpl implements CleanerService {
 
     @Override
     public Result update(CleanerEntityURequest cleanerEntityURequest) {
+        Optional<CleanerEntity> optionalCleanerEntity=this.cleanerEntityRepository.findById(cleanerEntityURequest.getId());
+        if(!optionalCleanerEntity.isPresent()){
+            return new ErrorResult("cleaner bulunamadı.");
+        }
         this.cleanerEntityRepository.save(cleanerMapper.cleanerUpdateToEntity(cleanerEntityURequest));
         return new SuccessResult("güncellendi");
     }
 
     @Override
     public Result delete(CleanerEntityDRequest cleanerEntityDRequest) {
+        Optional<CleanerEntity> cleanerEntity=this.cleanerEntityRepository.findById(cleanerEntityDRequest.getId());
+        if(!cleanerEntity.isPresent()){
+            return new ErrorResult("cleaner bulunamadı.");
+        }
        this.cleanerEntityRepository.deleteById(cleanerEntityDRequest.getId());
        return new SuccessResult("silindi.");
     }
 
     @Override
     public DataResult<CleanerEntityDto> getById(int id) {
+        if(!this.cleanerEntityRepository.existsById(id)){
+            return new ErrorDataResult<>(null);
+        }
        CleanerEntity cleanerEntity=this.cleanerEntityRepository.getById(id);
        CleanerEntityDto cleanerEntityDto=cleanerMapper.cleanerEntityToDto(cleanerEntity);
        return new SuccessDataResult<>(cleanerEntityDto);
