+++ src/main/java/com/example/hospitalcleaner/business/service/impl/FeedbackServiceImpl.java
@@ -1,19 +1,18 @@
 package com.example.hospitalcleaner.business.service.impl;
 
 import com.example.hospitalcleaner.business.core.mapper.FeedbackMapper;
-import com.example.hospitalcleaner.business.core.results.DataResult;
-import com.example.hospitalcleaner.business.core.results.Result;
-import com.example.hospitalcleaner.business.core.results.SuccessDataResult;
-import com.example.hospitalcleaner.business.core.results.SuccessResult;
+import com.example.hospitalcleaner.business.core.results.*;
 import com.example.hospitalcleaner.business.dto.CleanerEntityDto;
 import com.example.hospitalcleaner.business.dto.FeedbackEntityDto;
 import com.example.hospitalcleaner.business.requests.*;
 import com.example.hospitalcleaner.business.service.FeedbackService;
 import com.example.hospitalcleaner.dataAccess.FeedbackEntityRepository;
+import com.example.hospitalcleaner.entity.FeedbackEntity;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class FeedbackServiceImpl implements FeedbackService {
@@ -37,12 +36,20 @@ public class FeedbackServiceImpl implements FeedbackService {
 
     @Override
     public Result update(FeedbackEntityURequest feedbackEntityURequest) {
+        Optional<FeedbackEntity> optionalFeedbackEntity=feedbackEntityRepository.findById(feedbackEntityURequest.getId());
+        if(!optionalFeedbackEntity.isPresent()){
+            return new ErrorResult("feedback bulunamadı.");
+        }
         this.feedbackEntityRepository.save(this.feedbackMapper.feedbackUpdateToEntity(feedbackEntityURequest));
         return new SuccessResult("güncellendi");
     }
 
     @Override
     public Result delete(FeedbackEntityDRequest feedbackEntityDRequest) {
+        Optional<FeedbackEntity> optionalFeedbackEntity=feedbackEntityRepository.findById(feedbackEntityDRequest.getId());
+        if(!optionalFeedbackEntity.isPresent()){
+            return new ErrorResult("feedback bulunamadı.");
+        }
         this.feedbackEntityRepository.deleteById(feedbackEntityDRequest.getId());
         return new SuccessResult("silindi.");
     }
